% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.findPeaks}
\alias{chromatogram.findPeaks}
\title{chromatogram.findPeaks}
\usage{
chromatogram.findPeaks(
  data,
  index = 1,
  id = NA,
  chromatogramDetectPeaksMethod = NA,
  ...
)
}
\arguments{
\item{data}{msInfo object containing the chromatogram with index ...  or id ...}

\item{index}{index of the chromatogram on which to do peak detection}

\item{id}{id of the chromatogram to perform peak detection on, overrides index}

\item{chromatogramDetectPeaksMethod}{function to use for peak detection. If NA (default)
then the function will use \link[MS.Analysis]{chromatogramFindPeaks}}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{chromatogramFindPeaks}
(or other function(s)) which does the actual peak detection}
}
\value{
a data.frame or NA
}
\description{
performs peak detection on chromatogram present in an msInfo object.
Can be used stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
\examples{
msR <- msInfo$new()
demoFile <- fs::path_package("extdata", "chrom1.csv", package = "MS.Analysis")
result <- read.table(demoFile, sep = ",", header = TRUE)
msR$chromatograms$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
chromatogram.findPeaks(data = msR$chromatograms, id = 1, signalNoiseRatio = 5)

}
