% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.intensity.max}
\alias{chromatogram.intensity.max}
\title{chromatogram.intensity.max}
\usage{
chromatogram.intensity.max(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the chromatogram with index ...  or id ...}

\item{index}{index of the chromatogram}

\item{id}{id of the chromatogram, overrides index}

\item{...}{to pass additional arguments on to the function. Redundant in this
specific function.}
}
\value{
numeric vector
}
\description{
finds the maximum intensity in a chromatogram present in an msInfo object.
Can be used stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
\examples{
msR <- msInfo$new()
demoFile <- fs::path_package("extdata", "chrom1.csv", package = "MS.Analysis")
result <- read.table(demoFile, sep = ",", header = TRUE)
msR$chromatograms$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
chromatogram.intensity.max(
  data = msR$chromatograms,
  id = 1,
  signalNoiseRatio = 5
)

}
