% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.plot.overlay}
\alias{chromatogram.plot.overlay}
\title{chromatogram.plot.overlay}
\usage{
chromatogram.plot.overlay(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the chromatograms with index ...  or id ...}

\item{index}{index of the chromatograms to be plotted}

\item{id}{id of the chromatograms to be plotted, overrides index}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{plotChromatogramOverlay}
which does the actual plotting}
}
\value{
a ggplot object or NA
}
\description{
plots an overlay of chromatograms present in an msInfo object.
Can be used stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
\examples{
msR <- msInfo$new()
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
msR$chromatograms$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
demoFile <- fs::path_package("extdata", "chrom1.csv", package = "MS.Analysis")
result <- read.table(demoFile, sep = ",", header = TRUE)
msR$chromatograms$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
chromatogram.plot.overlay(
  data = msR$chromatograms,
  id = 1:2,
  chromatogramColors = c("blue", "red")
)
}
