% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.plot}
\alias{chromatogram.plot}
\title{chromatogram.plot}
\usage{
chromatogram.plot(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the chromatogram with index ...  or id ...}

\item{index}{index of the chromatogram to be plotted}

\item{id}{id of the chromatogram to be plotted, overrides index}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{plotChromatogram}
which does the actual plotting.}
}
\value{
a ggplot object or NA
}
\description{
plots a chromatogram present in an msInfo object. Can be used
stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
\examples{
msR <- msInfo$new()
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
msR$chromatograms$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
chromatogram.plot(data = msR$chromatograms, id = 1)

}
