% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.plot.sum}
\alias{chromatogram.plot.sum}
\title{chromatogram.plot.sum}
\usage{
chromatogram.plot.sum(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the chromatograms with index ...  or id ...}

\item{index}{index of the chromatograms to be used}

\item{id}{id of the chromatograms to be used, overrides index}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{plotChromatogramSum}
which does the actual plotting}
}
\value{
a ggplot object or NA
}
\description{
sums the selected chromatograms and plots the area, while one
chromatogram is used as a reference of sorts. Can be used stand-alone or as
part of a \link[dataInfo]{dataFunctions} object.
}
