% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{chromatogram.smooth}
\alias{chromatogram.smooth}
\title{chromatogram.smooth}
\usage{
chromatogram.smooth(
  data,
  index = 1,
  id = NA,
  smoothFunction = smoothFunction.loess(x = "rt"),
  returnInfo = FALSE,
  ...
)
}
\arguments{
\item{data}{msInfo object containing the chromatogram with index ...  or id ...}

\item{index}{index of the chromatogram on which to do smoothing}

\item{id}{id of the chromatogram to perform smoothing on, overrides index}

\item{smoothFunction}{function to use for smoothing. Default is \link[MS.Analysis]{smoothFunction.loess}}

\item{returnInfo}{if the data is to be returned as a function that generates
a dataElement object (which can be used to put in an dataInfo object) via
the 'readData' function)}

\item{...}{to pass additional arguments on to smoothing function}
}
\value{
a data.frame of the smoothed data or a dataElement object
}
\description{
smooths a chromatogram present in an msInfo object.
Can be used stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
