% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_annotation.R
\name{chromatogramAnnotation}
\alias{chromatogramAnnotation}
\title{chromatogramAnnotation}
\description{
R6 Class dealing with mass spectrometry annotation data in an organised manner.
This class is a descendant of annotation
}
\note{
use with care, no safeguards (yet)
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
plotChromatogram(chromatogram = result)
thePeaks <- chromatogramFindPeaks(result, span = 2, smoothing = NA, signalNoiseRatio = 10)
thePeaks |> head()
theAnnotation <- chromatogramAnnotation$new(
  rts = thePeaks$peak_rt,
  labels = formatDigits(2)(thePeaks$peak_rt)
)
theAnnotation
plotChromatogram(chromatogram = result, annotateRt = list(theAnnotation))
plotChromatogram(chromatogram = result, rtLimits = c(7, 11), annotateRt = list(theAnnotation),
                 incrScaleIntensity = 0.35)
theAnnotation$check(chromatogram = result)
theAnnotation
theAnnotation$labelType <- 2
plotChromatogram(chromatogram = result, rtLimits = c(7, 11), annotateRt = list(theAnnotation),
                 incrScaleIntensity = 0.35)
theAnnotation$labelType <- 3
plotChromatogram(chromatogram = result, rtLimits = c(7, 11), annotateRt = list(theAnnotation),
                 incrScaleIntensity = 0.35)

}
\section{Super class}{
\code{\link[MS.Analysis:annotation]{annotation}} -> \code{chromatogramAnnotation}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{rt}}{gets/sets the x values of the labels}

    \item{\code{intensity}}{gets/sets the y-positions in the object}

    \item{\code{labelType}}{gets/sets the axisConnectType
When set it takes care of axisConnect & labelConnect at the same time}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-chromatogramAnnotation-initialize}{\code{chromatogramAnnotation$new()}}
    \item \href{#method-chromatogramAnnotation-check}{\code{chromatogramAnnotation$check()}}
    \item \href{#method-chromatogramAnnotation-clone}{\code{chromatogramAnnotation$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="MS.Analysis" data-topic="annotation" data-id="draw"><a href='../../MS.Analysis/html/annotation.html#method-annotation-draw'><code>annotation$draw()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="MS.Analysis" data-topic="annotation" data-id="print"><a href='../../MS.Analysis/html/annotation.html#method-annotation-print'><code>annotation$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chromatogramAnnotation-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-chromatogramAnnotation-initialize}{}}}
\subsection{\code{chromatogramAnnotation$new()}}{
  Initializes an spectrum annotation object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{chromatogramAnnotation$new(
  annotationName = NA,
  nameX = "rt",
  nameY = "intensity",
  rts = NA,
  labels = NA,
  levelWhere = 0.9,
  levelWhereAbsolute = FALSE,
  labelColor = "red",
  labelSize = 1,
  labelAngle = 30,
  labelConnect = FALSE,
  labelConnectColor = "red",
  labelConnectAlpha = 0.25,
  labelBetween = FALSE,
  labelNudgeY = TRUE,
  axisConnect = TRUE,
  axisConnectColor = "red",
  axisConnectAlpha = 0.25,
  axisConnectWidth = 0.5,
  axisConnectLevel = NA,
  axisConnectWhere = 0.05,
  axisConnectType = 1
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{annotationName}}{set the internal name of the object}
      \item{\code{nameX}}{sets the name of the x-axis variable (eg m/z or retention time)}
      \item{\code{nameY}}{sets the name of the y-axis variable (eg intensity)}
      \item{\code{rts}}{rt-values (retention time, location on x-axis) for the labels}
      \item{\code{labels}}{the labels to be used for the annotation. Should have same length
as argument rts. If NA, then it will be as.character(rts)}
      \item{\code{levelWhere}}{y-axis height where axis labels are to be placed. Expressed as
a fraction of the y-axis maximum (default 0.9)}
      \item{\code{levelWhereAbsolute}}{labelWhere positions are absolute? (default: FALSE)}
      \item{\code{labelColor}}{sets the color of the labels}
      \item{\code{labelSize}}{sets the size of the labels}
      \item{\code{labelAngle}}{sets the angle of display of the labels}
      \item{\code{labelConnect}}{logical vector which defines if the labels are connected
horizontally ('along' the x-axis)}
      \item{\code{labelConnectColor}}{sets the color of the connection lines between the labels}
      \item{\code{labelConnectAlpha}}{sets the alpha of the connection lines between the labels}
      \item{\code{labelBetween}}{logical vector. If TRUE then labels are placed between connection points}
      \item{\code{labelNudgeY}}{numeric vector, to have the labels drawn a little higher or lower in the final graph}
      \item{\code{axisConnect}}{specifies if the labels should have axis connect lines}
      \item{\code{axisConnectColor}}{sets the color of the axis connector line(s)}
      \item{\code{axisConnectAlpha}}{sets the alpha of the axis connector line(s)}
      \item{\code{axisConnectWidth}}{sets the width of the axis connector line(s)}
      \item{\code{axisConnectLevel}}{defines to where (how far down) the axisConnect lines should reach}
      \item{\code{axisConnectWhere}}{this is how high above the connect, is relative to max Y-axis.
To place labels close to their peaks, this option should only be set/changed by
annotation field labelType}
      \item{\code{axisConnectType}}{at this moment can be 1,2 or 3. 1 means connect all the way down,
2 means up till axisConnectLevel above int of m/z. 3 = overrides all axisConnect Settings,
labels put at axisConnectLevel above int of m/z and axisConnect lines}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a new spectrum annotation object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chromatogramAnnotation-check"></a>}}
\if{latex}{\out{\hypertarget{method-chromatogramAnnotation-check}{}}}
\subsection{\code{chromatogramAnnotation$check()}}{
  checks whether the m/z coordinates have 'sufficient' intensity labels
if not, then the label is dropped from the object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{chromatogramAnnotation$check(
  chromatogram,
  toleranceLow = 0.1,
  toleranceHigh = toleranceLow,
  relativeCutOff = TRUE,
  intensityCutOff = ifelse(relativeCutOff, 0.001, 10)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{chromatogram}}{data.frame with at least an m/z column and an intensity column}
      \item{\code{toleranceLow}}{left tolerance (m/z)}
      \item{\code{toleranceHigh}}{right tolerance (m/z)}
      \item{\code{relativeCutOff}}{is the intensity CutOff reletive (fraction of maximum intensity)}
      \item{\code{intensityCutOff}}{y-axis (intensity) cut off y-axis value (intensity) should be above}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    the object itself (invisible)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chromatogramAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-chromatogramAnnotation-clone}{}}}
\subsection{\code{chromatogramAnnotation$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{chromatogramAnnotation$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
