% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_functional.R
\name{chromatogramPeaks.FixedEdges}
\alias{chromatogramPeaks.FixedEdges}
\title{chromatogramPeaks.FixedEdges}
\usage{
chromatogramPeaks.FixedEdges(peaks, edges = c(0.075, 0.15))
}
\arguments{
\item{peaks}{peak table data.frame as coming from chromatogramFindPeaks function.
The data.frame needs to have at least the peak_rt column (which shows the retention
time of the peak)}

\item{edges}{2-element numeric vector specifying the lower limit and the upper limit
for the peak area calculation (other function). They're relative numbers:
peak_rt+edges til peak_rt$edges}
}
\value{
the peak table data.frame with left/right columns
}
\description{
set fixed left, right limits in the peak table. Rather simple
convenience function. Can be used for simple peak area determination where
chromatographic peaks are relatively far away from each other
}
\note{
The left/right values are needed as limits to calculate the peak area
(in another function). The method used in this function is obviously quite
crude and will not be appropiate for a lot of (more complex) peaks/chromatograms.
For those we need more advanced ways of determining the beginning and end of
a chromatographic peak
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
result |> head()
plot(result, type = "l")
peaks <- data.frame(
  peak_rt = c(11.074, 12.133, 17.769, 19.515),
  peak_intensity = c(145412448, 151731744, 99109632, 70294472),
  left = NA,
  right = NA,
  area = NA
)
peaks
peaks <- peaks |> chromatogramPeaks.FixedEdges()
peaks
plot(result, type = "l", xlim = c(10, 13), ylim = c(0, 2E8))
text(
  peaks$peak_rt[1:2],
  peaks$peak_intensity[1:2],
  formatMinimumDigits(1)(peaks$peak_rt[1:2]),
  cex = 0.65,
  pos = 3,
  col = "red"
)
abline(
  v = peaks$peak_rt[1:2],
  col = scales::alpha(rgb(1, 0, 0), 0.75),
  lty = "dotted"
)
abline(
  v = peaks$left[1:2],
  col = scales::alpha(rgb(0, 0, 1), 0.75),
  lty = "dashed"
)
abline(
  v = peaks$right[1:2],
  col = scales::alpha(rgb(0, 0, 1), 0.75),
  lty = "dashed"
)

}
