% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_annotation.R
\name{createIonSeries}
\alias{createIonSeries}
\title{createIonSeries}
\usage{
createIonSeries(
  whichSeries = "y",
  theSeries = NA,
  whichCharge = 1,
  positive = TRUE,
  chargeConnectorPre = "",
  chargeConnector = "",
  chargeConnectorPost = "",
  seriesConnectorPre = "_",
  seriesConnectorPost = ""
)
}
\arguments{
\item{whichSeries}{letter or name assigned to the series, eg y for y-ions,
b for b-ions, etc}

\item{theSeries}{integer vector, eg 1:10 to specify the numbers of the ion series}

\item{whichCharge}{integer vector, gives the charge state (1,2,3,...)}

\item{positive}{logical vector. If TRUE than positive ions series is generated,
else negative ion series}

\item{chargeConnectorPre}{character vector placed between the charge and the
chargeConnector}

\item{chargeConnector}{character vector to connect the name + charge to the
ion series number}

\item{chargeConnectorPost}{character vector placed between the chargeConnector
and the ion series number}

\item{seriesConnectorPre}{character vector placed just before the series
number}

\item{seriesConnectorPost}{character vector placed just after the series
number}
}
\value{
a character vector or NA
}
\description{
function to generate a series of labels for ions
}
\note{
all arguments, except where indicated, are character vectors
}
\examples{
createIonSeries(theSeries = 1:5)
createIonSeries(whichSeries = "b", theSeries = 1:5, whichCharge = 2)

}
