% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{fileInfo.CSV}
\alias{fileInfo.CSV}
\title{fileInfo.CSV}
\usage{
fileInfo.CSV(filename)
}
\arguments{
\item{filename}{name of the file from which the info is to be read. This must
be a csv-style file that has headers and can be read using
\link[utils]{read.csv}}
}
\value{
a function that returns list of two objects: info (contains filename
and description) and data (empty)
}
\description{
generates a function for a simple info element for a CSV function.
Note that it does not read data, it only generates an object with the filename
and the names of the columns in the info object
}
\note{
the returned function has two (functional) arguments: sep (default = ",")
and ... which can be used to pass additional arguments onto
\link[utils]{read.csv}
}
\examples{
filename <- tempfile()
utils::write.csv(datasets::mtcars, file = filename, row.names = FALSE)
result <- fileInfo.CSV(filename = filename)()
result[[1]]$info
result[[1]]$data

}
