% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{formatDigits}
\alias{formatDigits}
\title{formatDigits}
\usage{
formatDigits(digits)
}
\arguments{
\item{digits}{integer value that specifies the number of digits to be used
by the resulting function}
}
\value{
a function that will take a numeric vector as an argument and
returns a character vector of the numeric vector with the set number of
digits (see ?scales::lebel_number for more info)
}
\description{
Function factory to be used to specify the number of digits to be
used in numbers
}
\note{
this is more or less an example of a function to be used to specify
axis-label formats in ggplot2.
}
\examples{
formatDigits(0)(pi)
formatDigits(1)(pi)
formatDigits(2)(pi)
formatDigits(5)(pi)
singleDigit <- formatDigits(2)
singleDigit(pi)

}
