% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{formatDigitsWaiver}
\alias{formatDigitsWaiver}
\title{formatDigitsWaiver}
\usage{
formatDigitsWaiver(v)
}
\arguments{
\item{v}{numeric vector to be converted to a character vector}
}
\value{
a character vector
}
\description{
Function to convert numerical vectors into strings. This specific one can be
used in stead of the function factories below if needed when no options are needed
}
\examples{
formatDigitsWaiver(1.234)
formatDigitsWaiver(1)
formatDigitsWaiver(1.234E01)
formatDigitsWaiver(1.234E-01)
formatDigitsWaiver(pi)

}
