% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{formatScientificDigits}
\alias{formatScientificDigits}
\title{formatScientificDigits}
\usage{
formatScientificDigits(digits)
}
\arguments{
\item{digits}{integer value that specifies the number of digits to be used
by the resulting function}
}
\value{
a function that will take a numeric vector as an argument and
returns a character vector of the numeric vector in scientific format with
the set number of digits (see ?scales::lebel_scientific for more info)
}
\description{
Function factory to be used to specify the number of digits to be
used in numbers when using scientific notation
}
\examples{
formatScientificDigits(4)(pi)
formatScientificDigits(10)(pi)
formatScientificDigits(3)(1.4556E11)

}
