% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{formatScientificMinimumDigits}
\alias{formatScientificMinimumDigits}
\title{formatScientificMinimumDigits}
\usage{
formatScientificMinimumDigits(digits)
}
\arguments{
\item{digits}{integer value that specifies the number of digits to be used
by the resulting function}
}
\value{
a function that will take a numeric vector as an argument and
returns a character vector of the numeric vector in scientific format with
the set number of digits (see ?formatC for more info)
}
\description{
Function factory to be used to specify the minimum number of digits
to be used in numbers when using scientific notation
}
\examples{
formatScientificMinimumDigits(5)(1.4556E11)
formatScientificMinimumDigits(15)(pi)

}
