% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.Class}
\alias{is.Class}
\title{is.Class}
\usage{
is.Class(object, whichClass)
}
\arguments{
\item{object}{a data object of some class}

\item{whichClass}{character string: class name to be tested}
}
\value{
TRUE or FALSE
}
\description{
internal helper function to determine if object == whichClass or
a descendant of a data type
}
\examples{
is.Class(c("A", "C"), "character")
is.Class(c("A", "C"), "integer")
is.Class(1:4, "character")
is.Class(1:4, "integer")

}
