% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatograms_graphics.R
\name{plotChromatogram}
\alias{plotChromatogram}
\title{plotChromatogram}
\usage{
plotChromatogram(
  chromatogram,
  rtLimits = NULL,
  intensityLimits = NULL,
  incrScaleIntensity = 0.05,
  scaleIntensityLocal = TRUE,
  intensityPercentage = FALSE,
  rtLabelFormat = formatDigits(1),
  intensityLabelFormat = ifelse(intensityPercentage, formatDigits(0),
    formatScientificDigits(4)),
  lineType = "solid",
  lineAlpha = 1,
  lineColor = "black",
  lineWidth = NA,
  areaFill = NA,
  areaAlpha = 0.5,
  rtUnits = "(mins)",
  rtTitle = paste0("rt ", rtUnits),
  intensityTitle = ifelse(intensityPercentage, "Intensity (\%)", "Intensity"),
  labels = NA,
  labelsLabel = "peak_rt",
  labelsLabelFormat = ifelse(labelsLabel == "peak_rt", formatDigits(2),
    ifelse(labelsLabel == "area", formatScientificDigits(4), ifelse(labelsLabel ==
    "peak_intensity", formatScientificDigits(3), formatDigitsWaiver))),
  labelCutOffValue = ifelse(labelsLabel == "peak_rt", "peak_intensity", labelsLabel),
  labelCutOff = 0.01,
  labelOverlap = FALSE,
  labelSize = 3,
  labelColor = "black",
  labelAngle = 25,
  labelNudge_x = ifelse(labelsLabel == "peak_rt", 0.01, 0.02),
  labelNudge_y = ifelse(labelsLabel == "peak_rt", 0.02, 0.025),
  annotateRt = NULL,
  chromatogramTitle = ggplot2::waiver(),
  chromatogramSubtitle = ggplot2::waiver(),
  chromatogramCaption = ggplot2::waiver(),
  plot.margins.default = FALSE,
  plot.margins = c(5, 15, 5, 5),
  plot.margins.units = "points",
  generalTextSize = NA,
  generalLineWidth = NA,
  overrideLineWidth = FALSE
)
}
\arguments{
\item{chromatogram}{a data.frame with "rt" & "intensity" columns}

\item{rtLimits}{two element numeric vector. Essentially the limits of the x-axis
(rt, retention time)}

\item{intensityLimits}{two element numeric vector. Essentially the limits of the
y-axis (intensity)}

\item{incrScaleIntensity}{numeric value which specifies the factor with which
to increase the y-axis limit}

\item{scaleIntensityLocal}{logical vector which determines whether to use global
intensities (FALSE) or only the intensities of m/z's which fall within the mzLimits
(TRUE, default)}

\item{intensityPercentage}{Whether to display the intensity axis in percentages
(default FALSE)}

\item{rtLabelFormat}{defines the format of the rt (x) axis labels.
See eg \link[MS.Analysis]{formatDigits}. Default is \link[ggplot2]{waiver}
which ensures 'standard' formatting}

\item{intensityLabelFormat}{defines the format of the intensity (y) axis labels.
See eg \link[MS.Analysis]{formatDigits}}

\item{lineType}{determines linetype of the lines used for plotting the chromatogram}

\item{lineAlpha}{determines alpha of the lines used for plotting the chromatogram}

\item{lineColor}{determines color of the lines used for plotting the chromatogram}

\item{lineWidth}{determines linewidth of the lines used for plotting the chromatogram}

\item{areaFill}{default is NA. If not NA, then should be color that the area
under the chromatogram is given. See example.}

\item{areaAlpha}{the alpha of the filled area under the chromatogram, default
is 0.5}

\item{rtUnits}{together with rtTitle argument: label for x-axis}

\item{rtTitle}{together with rtUnits argument: label for x-axis}

\item{intensityTitle}{defines the label for y-axis (intensity)}

\item{labels}{supposed to be a data.frame like coming from
\link[MS.Analysis]{chromatogramFindPeaks}. This may change in the future}

\item{labelsLabel}{defines which column of the labels data.frame to use as
label (default is 'peak_rt')}

\item{labelsLabelFormat}{defines the format of the labels}

\item{labelCutOffValue}{defines which column of the labels data.frame to use
for the labelCutOff argument}

\item{labelCutOff}{numeric vector that defines at which intensity level, peak
labels should not be shown. Default is 0.01 which means that below 1\% of
maximum intensity the labels won't be shown. Note: this assumes the column
chosen by labelCutOffValue is intensity related}

\item{labelOverlap}{logical vector that sets the "check_overlap" argument of
\link[ggplot2]{geom_text} which is used internally for labels}

\item{labelSize}{sets the size of the label text}

\item{labelColor}{sets the color of the label text}

\item{labelAngle}{sets the angle of the label text}

\item{labelNudge_x}{determines how much the label should be shifted (nudged)
horizontally}

\item{labelNudge_y}{determines how much the label should be shifted (nudged)
vertically}

\item{annotateRt}{either NA or a list of annotation objects to be used, see
\link[MS.Analysis]{chromatogramAnnotation}}

\item{chromatogramTitle}{sets the title of the chromatogram}

\item{chromatogramSubtitle}{sets the subtitle of the chromatogram}

\item{chromatogramCaption}{sets the caption of the chromatogram}

\item{plot.margins.default}{sets whether default ggplot margins are to be used
(default = FALSE)}

\item{plot.margins}{sets the plot margin sizes/widths (4 element integer vector).
Default is c(5,15,5,5)}

\item{plot.margins.units}{sets the units for the margin sizes/widths. Default
is "point"}

\item{generalTextSize}{factor with which to increase or decrease the size of
all text elements used. Default is NA}

\item{generalLineWidth}{factor with which to increase or decrease the width of
all line elements used. Default is NA}

\item{overrideLineWidth}{if the lineWidth argument is not used (NA), then if
this argument is TRUE line width 0.5 x generalLineWidth is used, if FALSE then
the line width will be set to 0.25 x generalLineWidth}
}
\value{
a ggplot object
}
\description{
Creates a plot of a chromatogram, rt (retention time) will be on
the x-axis, intensity on the y-axis
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
plotChromatogram(chromatogram = result)
plotChromatogram(chromatogram = result, intensityPercentage = TRUE)
plotChromatogram(chromatogram = result, rtLimits = c(3, 7))
plotChromatogram(
  chromatogram = result,
  labels = chromatogramFindPeaks.General(
    smoothing = NA,
    span = 1,
    signalNoiseRatio = 8
  )(result),
  labelColor = "red"
)
plotChromatogram(
  chromatogram = result,
  rtLimits = c(3, 7),
  labels = chromatogramFindPeaks.General(
    smoothing = NA,
    span = 1,
    signalNoiseRatio = 1
  )(result),
  labelColor = "red",
  lineColor = "blue"
)
plotChromatogram(
chromatogram = result,
rtLimits = c(3, 7),
areaFill = "red",
areaAlpha = 0.5
)

}
