% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatograms_graphics.R
\name{plotChromatogramOverlay}
\alias{plotChromatogramOverlay}
\title{plotChromatogramOverlay}
\usage{
plotChromatogramOverlay(
  chromatogramList,
  chromatogramColors = "black",
  chromatogramLineTypes = "solid",
  chromatogramAlphas = 1,
  chromatogramWidths = NA,
  chromatogramAreaFills = NA,
  chromatogramAreaAlphas = 0.5,
  rtLimits = NULL,
  intensityLimits = NULL,
  incrScaleIntensity = 0.05,
  scaleIntensityLocal = TRUE,
  intensityPercentage = FALSE,
  rtLabelFormat = formatDigits(1),
  intensityLabelFormat = ifelse(intensityPercentage, formatDigits(0),
    formatScientificDigits(4)),
  annotateRt = NULL,
  rtUnits = "(mins)",
  rtTitle = paste0("rt ", rtUnits),
  intensityTitle = ifelse(intensityPercentage, "Intensity (\%)", "Intensity"),
  chromatogramTitle = ggplot2::waiver(),
  chromatogramSubtitle = ggplot2::waiver(),
  chromatogramCaption = ggplot2::waiver(),
  plot.margins.default = FALSE,
  plot.margins = c(5, 15, 5, 5),
  plot.margins.units = "points",
  generalTextSize = NA,
  generalLineWidth = NA,
  overrideLineWidth = FALSE
)
}
\arguments{
\item{chromatogramList}{list of data.frame's (chromatograms)}

\item{chromatogramColors}{vector specifying the colors of the different chromatograms,
eg c("red","black"). Length 1 or same length as chromatogramList argument}

\item{chromatogramLineTypes}{vector specifying the linetypes of the different chromatograms.
Length 1 or same length as chromatogramList argument}

\item{chromatogramAlphas}{vector specifying the alphas of the different chromatograms.
Length 1 or same length as chromatogramList argument}

\item{chromatogramWidths}{vector specifying the linewidths of the different chromatograms.
Length 1 or same length as chromatogramList argument}

\item{chromatogramAreaFills}{vector specifying the area fill colors (for area
under the chromatograms) of the different chromatograms. Length 1 or same
length as chromatogramList argument under the chromatogram is given. See
example.}

\item{chromatogramAreaAlphas}{vector specifying the area fill alphas (for area
under the chromatograms) of the different chromatograms. Length 1 or same
length as chromatogramList argument under the chromatogram is given. See
example.}

\item{rtLimits}{two element numeric vector. Essentially the limits of the x-axis
(rt, retention time)}

\item{intensityLimits}{two element numeric vector. Essentially the limits of the
y-axis (intensity)}

\item{incrScaleIntensity}{numeric value which specifies the factor with which to
increase the y-axis limit}

\item{scaleIntensityLocal}{logical vector which determines whether to use global
intensities (FALSE) or only the intensities of m/z's which fall within the
rtLimits (TRUE, default)}

\item{intensityPercentage}{Whether to display the intensity axis in percentages
(default FALSE). Note: a 'side' effect is that all intensities are normalized
per chromatogram}

\item{rtLabelFormat}{defines the format of the rt (x) axis labels.
See eg \link[MS.Analysis]{formatDigits}. Default is \link[ggplot2]{waiver}
which ensures 'standard' formatting}

\item{intensityLabelFormat}{defines the format of the intensity (y) axis labels.
See eg \link[MS.Analysis]{formatDigits}}

\item{annotateRt}{either NA or a list of annotation objects to be used, see
\link[MS.Analysis]{chromatogramAnnotation}}

\item{rtUnits}{together with rtTitle argument: label for x-axis}

\item{rtTitle}{together with rtUnits argument: label for x-axis}

\item{intensityTitle}{defines the label for y-axis (intensity)}

\item{chromatogramTitle}{sets the title of the chromatogram}

\item{chromatogramSubtitle}{sets the subtitle of the chromatogram}

\item{chromatogramCaption}{sets the caption of the chromatogram}

\item{plot.margins.default}{sets whether default ggplot margins are to be used
(default = FALSE)}

\item{plot.margins}{sets the plot margin sizes/widths (4 element integer vector).
Default is c(5,15,5,5)}

\item{plot.margins.units}{sets the units for the margin sizes/widths. Default
is "point"}

\item{generalTextSize}{factor with which to increase or decrease the size of
all text elements used. Default is NA}

\item{generalLineWidth}{factor with which to increase or decrease the width of
all line elements used. Default is NA}

\item{overrideLineWidth}{if the lineWidth argument is not used (NA), then if
this argument is TRUE line width 0.5 x generalLineWidth is used, if FALSE then
the line width will be set to 0.25 x generalLineWidth}
}
\value{
a ggplot object
}
\description{
Creates a overlay plot of a list of chromatograms, rt (retention
time) will be on the x-axis, intensity on the y-axis.
}
\examples{
demoFile <- fs::path_package("extdata", "Data0001.CSV", package = "MS.Analysis")
result <- readLines(demoFile, n = 9092)
result <- read.table(text = result, sep = ",", header = FALSE)[, 2:3]
colnames(result) <- c("rt", "intensity")
result2 <- result
result2$rt <- result2$rt + 1
result2$intensity <- result2$intensity * 1.25
plotChromatogramOverlay(
  list(result, result2),
  chromatogramColors = c("black", "red")
)
plotChromatogramOverlay(
  list(result, result2),
  chromatogramColors = c("black", "red"),
  rtLimits = c(3, 7)
)
plotChromatogramOverlay(
  list(result, result2),
  chromatogramColors = c("black", "red"),
  chromatogramAreaFills = c("black", "red"),
  rtLimits = c(3, 7)
)


}
