% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_graphics.R
\name{plotSpectrumOverlay}
\alias{plotSpectrumOverlay}
\title{plotSpectrumOverlay}
\usage{
plotSpectrumOverlay(
  spectrumList,
  spectrumColors = "black",
  spectrumLineTypes = "solid",
  spectrumAlphas = 1,
  spectrumWidths = NA,
  mzLimits = NULL,
  intensityLimits = NULL,
  incrScaleIntensity = 0.05,
  scaleIntensityLocal = TRUE,
  intensityPercentage = FALSE,
  mzLabelFormat = ggplot2::waiver(),
  intensityLabelFormat = ifelse(intensityPercentage, formatDigits(0),
    formatScientificDigits(4)),
  centroidPlot = FALSE,
  cutOff = 0.01,
  annotateMz = NULL,
  plot.margins.default = FALSE,
  plot.margins = c(5, 15, 5, 5),
  plot.margins.units = "points",
  mzTitle = "m/z",
  intensityTitle = ifelse(intensityPercentage, "Intensity (\%)", "Intensity"),
  spectrumTitle = ggplot2::waiver(),
  spectrumSubtitle = ggplot2::waiver(),
  spectrumCaption = ggplot2::waiver(),
  generalTextSize = NA,
  generalLineWidth = NA,
  overrideLineWidth = FALSE
)
}
\arguments{
\item{spectrumList}{list of data.frame's (spectra)}

\item{spectrumColors}{vector specifying the colors of the different spectra,
eg c("red","black"). Length 1 or same length as spectrumList argument}

\item{spectrumLineTypes}{vector specifying the linetypes of the different spectra.
Length 1 or same length as spectrumList argument}

\item{spectrumAlphas}{vector specifying the alphas of the different spectra.
Length 1 or same length as spectrumList argument}

\item{spectrumWidths}{vector specifying the linewidths of the different spectra.
Length 1 or same length as spectrumList argument}

\item{mzLimits}{two element numeric vector. Essentially the limits of the x-axis
(m/z)}

\item{intensityLimits}{two element numeric vector. Essentially the limits of the
y-axis (intensity)}

\item{incrScaleIntensity}{numeric value which specifies the factor with which to
increase the y-axis limit}

\item{scaleIntensityLocal}{logical vector which determines whether to use global
intensities (FALSE) or only the intensities of m/z's which fall within the mzLimits
(TRUE, default)}

\item{intensityPercentage}{Whether to display the intensity axis in percentages
(default FALSE)}

\item{mzLabelFormat}{defines the format of the m/z (x) axis labels.
See eg \link[MS.Analysis]{formatDigits}. Default is \link[ggplot2]{waiver} which
ensures 'standard' formatting}

\item{intensityLabelFormat}{defines the format of the intensity (y) axis labels.
See eg \link[MS.Analysis]{formatDigits}}

\item{centroidPlot}{logical vector that defines whether the plot should be centroid
-like (\link[ggplot2]{geom_segment} is used) or profile-like (\link[ggplot2]{geom_line}
is used)}

\item{cutOff}{numeric vector that defines at which intensity level, below which
peaks should not be shown. Default is 0.01 which means that below 1\% of maximum
intensity the peaks won't be shown. Currently only works for centroid(ed) spectra.}

\item{annotateMz}{either NA or a list of annotation objects to be used, see
\link[MS.Analysis]{spectrumAnnotation}}

\item{plot.margins.default}{sets whether default ggplot margins are to be used
(default = FALSE)}

\item{plot.margins}{sets the plot margin sizes/widths (4 element integer vector).
Default is c(5,15,5,5)}

\item{plot.margins.units}{sets the units for the margin sizes/widths. Default
is "point"}

\item{mzTitle}{defines the label for the x-axis (m/z)}

\item{intensityTitle}{defines the label for y-axis (intensity)}

\item{spectrumTitle}{sets title of the spectrum}

\item{spectrumSubtitle}{sets subtitle of the spectrum}

\item{spectrumCaption}{sets caption of the spectrum}

\item{generalTextSize}{factor with which to increase or decrease the size of
all text elements used, except for annotateMz elements. Default is NA}

\item{generalLineWidth}{factor with which to increase or decrease the width of
all line elements used, except for annotateMz elements. Default is NA}

\item{overrideLineWidth}{if the lineWidth argument is not used (NA), then if
this argument is TRUE line width 0.5 x generalLineWidth is used, if FALSE then
the line width will be set to 0.25 x generalLineWidth}
}
\value{
a ggplot object
}
\description{
Creates a overlay plot of a list of spectra, m/z will be on the
x-axis, intensity on the y-axis
}
\examples{
demoFile <- fs::path_package("extdata", "spec2.csv", package = "MS.Analysis")
profileSpectrum <- read.table(demoFile, sep = ",", header = TRUE)

plotSpectrum(profileSpectrum, labelCutOff = 0.025)
profileSpectrum2 <- profileSpectrum
profileSpectrum2$mz <- profileSpectrum2$mz + 10
profileSpectrum2$intensity <- profileSpectrum2$intensity + 5E6
plotSpectrumOverlay(list(profileSpectrum, profileSpectrum2),
                    spectrumColors = c("black", "red"))
plotSpectrumOverlay(list(profileSpectrum, profileSpectrum2),
                    spectrumColors = c("black", "red"), mzLimits = c(500, 600))

}
