% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(
  dataFrame = NA,
  columns = ifelseProper(identical(dataFrame, NA), NA, 1:ncol(dataFrame)),
  columnNames = NA,
  rowNames = NA,
  info = list(source = "data")
)
}
\arguments{
\item{dataFrame}{data.frame containing the data}

\item{columns}{specifies which columns to take from the dataFrame. Can be a character or
an integer vector}

\item{columnNames}{specifies which column names to give to the data.frame. Should be
same length as the columns argument.}

\item{rowNames}{specifies which row names to give to the data.frame. Should be same
length as the number of rows in the dataFrame argument. If rowNames = NA (default) then
rownames of the dataFrame are used. If set to NULL, then existing row names will be
removed}

\item{info}{named list object that contains info to be put into the info data.frame of
the info objects. The intention is to put some relevant information there, like source
of the data, file names and other information which makes it possible to recreate the
data}
}
\value{
a function that returns a list of two objects: info and data
}
\description{
basic function factory which generates a function that returns
a function that itself returns the data in a data.frame plus (if added)
some info about the data.frame
}
\note{
the reason these readData functions are function factories is that (depending on
the data source) the actual reading in of data may take quite some time. readData
provides a base function that reads the data in and returns a list of an info data.frame
object with info on the data and a data data.frame object that contains the actual data.
The result can be used as a data source for info objects (and their descendants)
}
\examples{
result <- readData(dataFrame = datasets::mtcars)()
result$data
result$info
result <- readData(dataFrame = datasets::mtcars, info = list(source = "datasets",
                                                             name = "mtcars"))()
                                                             result$info
result <- readData(dataFrame = datasets::mtcars, columns = 1:2)()
result$data
result <- readData(dataFrame = datasets::mtcars, columns = 1:2,
                   columnNames = c("x","y"), rowNames = NULL)()
result[["data"]]

}
