% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readDataFrame}
\alias{readDataFrame}
\title{readDataFrame}
\usage{
readDataFrame(
  dataFrame = NA,
  columns = NA,
  columnNames = NA,
  rowNames = NA,
  emptyData = data.frame(data = "No Data"),
  info = list(source = "data")
)
}
\arguments{
\item{dataFrame}{a -list- of data.frame's (or NA)}

\item{columns}{specifies which columns to take from the dataFrame. Can be a character or
an integer vector. Please note that the specified columns HAVE to exist in all of the
data.frame's in the list (dataFrame argument)}

\item{columnNames}{specifies which column names to give to the data.frame. Should be
same length as the columns argument. Applied to ALL data.frame's in the list}

\item{rowNames}{specifies which row names to give to the data.frame. Should be same
length as the number of rows in the dataFrame argument. If rowNames = NA (default) then
rownames of the dataFrame are used. If set to NULL, then existing row names will be
removed. Except for the values NA and NULL, this argument can only be used when the
number of rows in the different data.frame's in the list is exactly the same,
otherwise there may be errors or unexpected effects}

\item{emptyData}{defines what needs to be used for data when an element of the data.frame
list is NA. Default is data.frame(data = "No Data" ). This argument is used when
data.frame's are expected for the result.}

\item{info}{named list object that contains info to be put into the info data.frame of
the info objects. The intention is to put some relevant information there, like source
of the data, file names and other information which makes it possible to recreate the
data. Again, this argument is applied to all elements in the data.frame list}
}
\value{
a function that will generate a list of elements which each two objects:
info and data
}
\description{
basic function factory which generates a function that returns
a organized list of readData functions from a list of data.frames's.
Essentially a list version of \link{readData}
}
\examples{
result <- readDataFrame(dataFrame = list(datasets::mtcars, datasets::iris))
result
result()
result <- readDataFrame(dataFrame = list(datasets::mtcars, datasets::iris))()
result[[1]]
result[[1]]$data |> head()
result[[2]]$data |> head()
result <- readDataFrame(dataFrame = list(datasets::mtcars, datasets::iris),
                        columns = 1:2, columnNames = c("x","y"),
                        rowNames = NULL)()
result[[1]]$data |> head()
result[[2]]$data |> head()

}
