% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_graphics.R
\name{reduceSpectrumLabelDF}
\alias{reduceSpectrumLabelDF}
\title{reduceSpectrumLabelDF}
\usage{
reduceSpectrumLabelDF(
  spectrumLabelDF,
  maxIntensity = TRUE,
  number = 1,
  mzCloseness = 1
)
}
\arguments{
\item{spectrumLabelDF}{data.frame with columns: mz, intensity & label}

\item{maxIntensity}{logical vector: if TRUE then the top 'number' of
intensities will be returned. If FALSE, then the numbers will be reduced based
on how close they are togther. The 'mzClosness' argument together with the
'number' argument determine how many rows will be left}

\item{number}{the 'number' of top intensity rows to be returned}

\item{mzCloseness}{determines how the mz's values are grouped after which the
m/z's with the highest value within the groups are returned}
}
\value{
data.frame (same structure as spectrumLabelDF)
}
\description{
function to reduce number of rows to the top or bottom set.
Specific for spectrum labels
}
\note{
may be removed from export in future
}
\examples{
# create spectrum label data.frame
demoSpec <- fs::path_package("extdata", "spec1.csv", package = "MS.Analysis")
specData <- utils::read.table(file = demoSpec, header = TRUE, sep = ",")
specData$label <- as.character(specData$mz)
specData |> head()
specData |>
 dplyr::filter(mz > 150, mz < 180) |>
 head(20)
reduceSpectrumLabelDF(
  specData,
  number = 1,
  maxIntensity = FALSE,
  mzCloseness = 1
) |>
  dplyr::filter(mz > 150, mz < 180) |>
  head(20)

reduceSpectrumLabelDF(
  specData,
  number = 1,
  maxIntensity = FALSE,
  mzCloseness = log10(2)
) |>
  dplyr::filter(mz > 150, mz < 180) |>
  head(20)


}
