% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_functional.R
\name{smoothFunction.loess}
\alias{smoothFunction.loess}
\title{smoothFunction.loess}
\usage{
smoothFunction.loess(
  x = "mz",
  y = "intensity",
  formula = paste(c(y, x), collapse = "~"),
  span = 0.1
)
}
\arguments{
\item{x}{character vector that defines defines one of two column names to be
used in the \link[stats]{loess} function}

\item{y}{character vector that defines defines the second of two column names
to be used in the \link[stats]{loess} function}

\item{formula}{the default is y~x, but can be transformed in anything allowed
by the \link[stats]{loess} function}

\item{span}{alpha parameter than controls the amount of smoothing (see
\link[stats]{loess} for more info). Note: cannot be too low}
}
\value{
a function that will smoothen the data.frame provided by the argument
data. Additional parameters for the smoothing can be passed on by ...
argument
}
\description{
function factory for generating loess smoothing functions to be
applied to data in a data.frame
}
\note{
This is more of an example function factory. There is a lot possible
}
\examples{
set.seed(1234)
df <- data.frame(
  x = seq(0, 10, by = 0.1)
)
df$y <- 1+5*sin(df$x)+rnorm(101)
plot(df, type = "l")
plot(smoothFunction.loess(x = "x", y = "y")(df), type = "l")
plot(smoothFunction.loess(x = "x", y = "y", span = 0.5)(df), type = "l")
plot(smoothFunction.loess(x = "x", y = "y")(df, degree = 1), type = "l")

}
