% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{spectrum.basepeak.intensity}
\alias{spectrum.basepeak.intensity}
\title{spectrum.basepeak.intensity}
\usage{
spectrum.basepeak.intensity(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the spectrum with index ...  or id ...}

\item{index}{index of the spectrum to be plotted}

\item{id}{id of the spectrum to be plotted, overrides index}

\item{...}{to pass additional arguments on to the function. Redundant in this
specific function.}
}
\value{
a numeric vector
}
\description{
gets the highest intensity position in a spectrum. The spectrum
is present in an msInfo object. More meant for centroided spectra, but will
work with profile data too. Can be used stand-alone or as part of a
\link[dataInfo]{dataFunctions} object.
}
