% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{spectrum.mz.intensity}
\alias{spectrum.mz.intensity}
\title{spectrum.mz.intensity}
\usage{
spectrum.mz.intensity(
  data,
  index = 1,
  id = NA,
  mz = NA,
  mzRange = NA,
  intensityFunction = sum,
  ...
)
}
\arguments{
\item{data}{msInfo object containing the spectrum with index ...  or id ...}

\item{index}{index of the spectrum to be plotted}

\item{id}{id of the spectrum to be plotted, overrides index}

\item{mz}{numeric vector: the m/z for which the intensity is to be retrieved}

\item{mzRange}{one or two element numeric vector. If NA then it will be set
at c(0,0). A single element vector will have it's value used for both parts
(... below and ... above the m/z value).}

\item{intensityFunction}{function that determines how to deal with a set of
intensities present in mz +/- mzRange. Default is sum; other possibilities
are e.g. mean, max, min, etc.}

\item{...}{to pass additional arguments on to the function. Redundant in this
specific function.}
}
\value{
a numeric vector
}
\description{
gets the intensity at a certain m/z position (or m/z range) in
a spectrum. The spectrum is present in an msInfo object. More meant for
centroided spectra, but will work with profile data too. Can be used
stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
