% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{spectrum.plot.mirror}
\alias{spectrum.plot.mirror}
\title{spectrum.plot.mirror}
\usage{
spectrum.plot.mirror(data, index = 1:2, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the spectrum with index ...  or id ...}

\item{index}{index of the spectrum to be plotted}

\item{id}{id of the spectrum to be plotted, overrides index}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{plotSpectrumMirror} which
does the actual plotting}
}
\value{
a ggplot object or NA
}
\description{
creates a mirror plot of spectra present in an msInfo object. Can be used
stand-alone or as part of a \link[dataInfo]{dataFunctions}
}
