% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{spectrum.plot.overlay}
\alias{spectrum.plot.overlay}
\title{spectrum.plot.overlay}
\usage{
spectrum.plot.overlay(data, index = 1:data$length, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the spectrum with index ...  or id ...}

\item{index}{index of the spectrum to be plotted}

\item{id}{id of the spectrum to be plotted, overrides index}

\item{...}{to pass additional arguments on to \link[MS.Analysis]{plotSpectrumOverlay} which
does the actual plotting}
}
\value{
a ggplot object or NA
}
\description{
plots an overlay of spectra present in an msInfo object. Can be used
stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
