% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msData.R
\name{spectrum.tic}
\alias{spectrum.tic}
\title{spectrum.tic}
\usage{
spectrum.tic(data, index = 1, id = NA, ...)
}
\arguments{
\item{data}{msInfo object containing the spectrum with index ...  or id ...}

\item{index}{index of the spectrum to be plotted}

\item{id}{id of the spectrum to be plotted, overrides index}

\item{...}{to pass additional arguments on to the function. Redundant in this
specific function.}
}
\value{
a numeric vector
}
\description{
calculates the total ion current (TIC) of a spectrum. It's nothing
more than a sum of the intensities present. The spectrum is present in an msInfo
object. Can be used stand-alone or as part of a \link[dataInfo]{dataFunctions} object.
}
\examples{
msR <- msInfo$new()
demoFile <- fs::path_package("extdata", "spec2.csv", package = "MS.Analysis")
result <- read.table(demoFile, sep = ",", header = TRUE)
msR$spectra$add(
  dataElements = dataInfo::readDataFrame(dataframes = list(result))
)
spectrum.tic(data = msR$spectra)

}
