% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_graphics.R
\name{spectrumLabels}
\alias{spectrumLabels}
\title{spectrumLabels}
\usage{
spectrumLabels(
  dataframe = NA,
  mz = NA,
  intensity = NA,
  label = dataInfo::formatDigits(4),
  top = NA,
  mzAccuracy = NA,
  maxIntensity = TRUE
)
}
\arguments{
\item{dataframe}{data.frame (with columns: mz, intensity, label). If this parameter
is not NA, it will override mz, intensity arguments}

\item{mz}{numeric vector specifying the m/z's of the labels}

\item{intensity}{numeric vector specifying the intensity of the labels}

\item{label}{function that specifies the formatting of the label, example is
formatDigits(4) (default)}

\item{top}{integer vector. If not NA only show labels with top highest
intensities}

\item{mzAccuracy}{integer vector, can be both positive and negative or NA. If
NA then this is not used. Values will group together 10^mzAccuracy values
and show only the highest one. Eg mzAccuracy NA: spectrumLabels(mz =
c(100,100.0001)), nothing will happen to labels; spectrumLabels(mz =
c(100,101.0001), mzAccuracy = 0) nothing will happen, but with value 1
only the highest intensity (or lowest m/z) will still be in data.frame
after the function. In case of spectrumLabels(mz = c(100,100.0001),
mzAccuracy = -4), both values will still be there, if value is 3 then only
100.0000 is there}

\item{maxIntensity}{if TRUE then the top 'number' of intensities will be
returned.}
}
\value{
data.frame (with columns: mz, intensity, label)
}
\description{
function to generate a data.frame of labels (each row is a
separate label). Specific for spectrum labels
}
\examples{
demoSpec <- fs::path_package("extdata", "spec1.csv", package = "MS.Analysis")
specData <- utils::read.table(file = demoSpec, header = TRUE, sep = ",")
spectrumLabels(specData)
spectrumLabels(specData, mzAccuracy = 1)
spectrumLabels(specData, mzAccuracy = 1, maxIntensity = TRUE, top = 10)

}
